30/08/2024
Working Util for 4.

+++ Lms/src/main/java/com/emo/matrix/lms/controller/CourseController.java
@@ -1,18 +1,24 @@
 package com.emo.matrix.lms.controller;
 
 import com.emo.matrix.lms.dto.CourseDTO;
+import com.emo.matrix.lms.exception.ResourceNotFoundException;
+import com.emo.matrix.lms.helper.FileUploadHelper;
 import com.emo.matrix.lms.models.Course;
+import com.emo.matrix.lms.repository.CourseRepository;
 import com.emo.matrix.lms.service.CourseService;
 import com.emo.matrix.lms.utils.CourseUtil;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/course")
@@ -22,32 +28,42 @@ public class CourseController {
     private CourseService courseService;
     
     @Autowired
-    private CourseUtil courseUtil;
+    private CourseRepository courseRepository;
+    
+//    @Autowired
+//    private CourseUtil courseUtil;
+    
+    @Autowired
+    private FileUploadHelper fileUploadHelper; 
 
     // Create a new course
     @PostMapping("/adminId/{adminId}/create")
     public ResponseEntity<CourseDTO> createCourse(@RequestBody Course course, @PathVariable Long adminId) {
     	Course craetedCourse = courseService.createCourse(course, adminId);
-        CourseDTO createdCourseDTO = courseUtil.toDTO(craetedCourse);
+        CourseDTO createdCourseDTO = CourseUtil.toDTO(craetedCourse);
         return ResponseEntity.ok(createdCourseDTO);
     }
 
     @PostMapping(value = "/{id}/upload", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE, "video/mp4"})
     public ResponseEntity<String> uploadVideo(@PathVariable Long id, @RequestParam("file") MultipartFile file) {
-//        try {
-//            Course course = courseService.uploadVideo(id, file);
-//            return ResponseEntity.ok("Video uploaded successfully!");
-//        } catch (RuntimeException e) {
-//            // Handle specific runtime exceptions
-//            return ResponseEntity.status(404).body(e.getMessage());
-//        } catch (Exception e) {
-//            // Handle general exceptions
-//            return ResponseEntity.status(500).body("Failed to upload video: " + e.getMessage());
-//        }
-    	System.out.println(file.getOriginalFilename());
-    	System.out.println(file.getSize());
-    	System.out.println(file.getContentType());
-    	System.out.println(file.getName());
+    	
+    	
+    	try {
+    		
+    		if(file.isEmpty()) {
+    			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Request must contain file");
+    		}
+    		
+    		boolean f = fileUploadHelper.uploadFile(file);
+    		if(f) {
+    			courseService.uploadVideo(id,file);
+    			
+//    			return ResponseEntity.ok(ServletUriComponentsBuilder.fromCurrentContextPath().path("/video/").path(file.getOriginalFilename()).toUriString());
+    			return ResponseEntity.ok("File is successfully uploaded");
+    		}
+    	} catch (Exception e) {
+    		e.printStackTrace();
+    	}
 
       return ResponseEntity.ok("Video uploaded successfully!");
     }
@@ -83,8 +99,9 @@ public class CourseController {
     // Get all courses
     @GetMapping("/")
     public ResponseEntity<List<CourseDTO>> getAllCourses() {
-        List<CourseDTO> courses = courseService.getAllCourses();
-        return ResponseEntity.ok(courses);
+    	List<Course> courses = courseService.getAllCourses();
+        List<CourseDTO> courseDTOs = courses.stream().map(CourseUtil::toDTO).collect(Collectors.toList());
+        return ResponseEntity.ok(courseDTOs);
     }
 
     // Get a course by ID
@@ -96,28 +113,34 @@ public class CourseController {
 
     // Delete a course by ID
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCourse(@PathVariable Long id) {
+    public ResponseEntity<String> deleteCourse(@PathVariable Long id) {
         try {
-            courseService.deleteCourse(id);
-            return ResponseEntity.noContent().build();
-        } catch (RuntimeException e) {
+        	if(courseRepository.existsById(id)) {
+        		courseService.deleteCourse(id);
+        		return ResponseEntity.ok("Cours with id " + id + " deleted successfully!.");
+        	}
+        	else {
+        		String errorMessage = "Course not found with id " + id;
+                return ResponseEntity.badRequest().body(errorMessage);
+        	}
+        } catch (ResourceNotFoundException e) {
             return ResponseEntity.notFound().build();
         }
     }
 
     // Get courses by a student's ID
-    @GetMapping("/student/{studentId}")
-    public ResponseEntity<List<CourseDTO>> getCoursesByStudentId(@PathVariable Long studentId) {
-        try {
-            List<CourseDTO> courses = courseService.getCoursesByStudentId(studentId);
-            return ResponseEntity.ok(courses);
-        } catch (RuntimeException e) {
-            return ResponseEntity.notFound().build();
-        }
-    }
+//    @GetMapping("/student/{studentId}")
+//    public ResponseEntity<List<CourseDTO>> getCoursesByStudentId(@PathVariable Long studentId) {
+//        try {
+//            List<CourseDTO> courses = courseService.getCoursesByStudentId(studentId);
+//            return ResponseEntity.ok(courses);
+//        } catch (RuntimeException e) {
+//            return ResponseEntity.notFound().build();
+//        }
+//    }
 
     // Enroll a student in a course
-    @PostMapping("/{courseId}/enroll/{studentId}")
+    @PostMapping("/{courseId}/enroll-student/{studentId}")
     public ResponseEntity<CourseDTO> enrollStudentInCourse(@PathVariable Long courseId, @PathVariable Long studentId) {
         try {
             CourseDTO course = courseService.enrollStudentInCourse(courseId, studentId);
@@ -128,11 +151,16 @@ public class CourseController {
     }
 
     // Remove a student from a course
-    @DeleteMapping("/{courseId}/remove/{studentId}")
-    public ResponseEntity<CourseDTO> removeStudentFromCourse(@PathVariable Long courseId, @PathVariable Long studentId) {
+    @DeleteMapping("/{courseId}/remove-student/{studentId}")
+    public ResponseEntity<String> removeStudentFromCourse(@PathVariable Long courseId, @PathVariable Long studentId) {
         try {
-            CourseDTO course = courseService.removeStudentFromCourse(courseId, studentId);
-            return ResponseEntity.ok(course);
+        	if(courseRepository.existsById(courseId)) {
+        		CourseDTO course = courseService.removeStudentFromCourse(courseId, studentId);
+        		return ResponseEntity.ok("Student account with id " + studentId + " removed from Course id "+ courseId +" successfully!.");
+        	} else {
+        		String errorMessage = "Course not found with id " + courseId;
+                return ResponseEntity.badRequest().body(errorMessage);
+        	}
         } catch (RuntimeException e) {
             return ResponseEntity.notFound().build();
         }
